feat(2fa): fix rebase mv 2fa to procedures
ref: MANAGER-14233
