+++ packages/manager/apps/procedures/src/data/hooks/useStatus.tsx
@@ -5,4 +5,5 @@ export const useFetch2faStatus = () =>
   useQuery({
     queryKey: ['get2faStatus'],
     queryFn: get2faStatus,
+    retry: 0,
   });
