+++ packages/manager/apps/procedures/src/pages/Home.page.test.tsx
@@ -8,8 +8,10 @@ const mockedUsedLocation = vi.fn();
 
 const fetch2faStatusFakeResponse = {
   data: { status: 'open' },
-  isFetched: true,
+  isFetched: false,
   isLoading: false,
+  isSuccess: false,
+  error: {},
 };
 
 vi.mock('react-router-dom', () => ({
@@ -29,6 +31,10 @@ vi.mock('@/data/hooks/useStatus', () => ({
 describe('Home.page', () => {
   it('should navigate to see page when status is open', async () => {
     fetch2faStatusFakeResponse.data.status = 'open';
+    fetch2faStatusFakeResponse.isLoading = false;
+    fetch2faStatusFakeResponse.isFetched = true;
+    fetch2faStatusFakeResponse.isSuccess = true;
+    fetch2faStatusFakeResponse.error = undefined;
 
     render(<Home />);
 
@@ -36,7 +42,15 @@ describe('Home.page', () => {
   });
 
   it('should navigate to create page when status is creationAuthorized', async () => {
-    fetch2faStatusFakeResponse.data.status = 'creationAuthorized';
+    fetch2faStatusFakeResponse.isFetched = true;
+    fetch2faStatusFakeResponse.error = {
+      response: {
+        status: 404,
+        data: {
+          class: 'Client::ErrNotFound::ErrNotFound',
+        },
+      },
+    };
 
     render(<Home />);
 
@@ -48,6 +62,8 @@ describe('Home.page', () => {
   it('should render outlet component when isFetched is true', async () => {
     fetch2faStatusFakeResponse.isLoading = false;
     fetch2faStatusFakeResponse.isFetched = true;
+    fetch2faStatusFakeResponse.isSuccess = true;
+    fetch2faStatusFakeResponse.error = undefined;
 
     render(<Home />);
