+++ packages/manager/apps/procedures/src/pages/create/form/Form.page.tsx
@@ -6,6 +6,7 @@ import {
 } from '@ovhcloud/ods-components';
 import {
   OsdsButton,
+  OsdsMessage,
   OsdsSelect,
   OsdsSelectOption,
   OsdsText,
@@ -20,6 +21,10 @@ import { FieldValues, useForm } from 'react-hook-form';
 import { FormDocumentFieldList } from './FormDocumentFields/FormDocumentFieldList';
 import { LegalFrom } from '@/types/user.type';
 import useUser from '@/context/User/useUser';
+import { useUploadDocuments } from '@/data/hooks/useDocments';
+import { ConfirmModal } from './Modal/ConfirmModal';
+import { SuccessModal } from './Modal/SuccessModal';
+import { ovhHomePageHref } from './contants/form.constants';
 
 const flatFiles = (files: FieldValues) =>
   Object.values(files)
@@ -33,9 +38,22 @@ const FormCreateRequest = () => {
 
   const { t } = useTranslation('account-disable-2fa');
   const { t: tdoc } = useTranslation('account-disable-2fa-documents');
-
   const [selectedByFRLegalForm, setSelectedByFRLegalForm] = useState<LegalFrom>(
     null,
+  );
+  const [showConfirmModal, setShowConfirmModal] = useState<boolean>(false);
+  const [showSuccessModal, setShowSuccessModal] = useState<boolean>(false);
+
+  const { mutate, isPending, isError, reset: resetUpload } = useUploadDocuments(
+    {
+      onSuccess: () => {
+        setShowSuccessModal(true);
+        setShowConfirmModal(false);
+      },
+      onError: () => {
+        setShowConfirmModal(false);
+      },
+    },
   );
 
   const files = flatFiles(watch());
@@ -45,6 +63,7 @@ const FormCreateRequest = () => {
     e: OsdsSelectCustomEvent<OdsSelectValueChangeEventDetail>,
   ) => {
     reset();
+    resetUpload();
     setSelectedByFRLegalForm(e.target.value as LegalFrom);
   };
 
@@ -58,7 +77,7 @@ const FormCreateRequest = () => {
   const { subsidiary } = user;
 
   return (
-    <form onSubmit={handleSubmit((data) => {})}>
+    <form onSubmit={handleSubmit(() => setShowConfirmModal(true))}>
       {isOtherLegalFormForFR && (
         <div className="my-6">
           <OsdsText
@@ -93,16 +112,33 @@ const FormCreateRequest = () => {
             subsidiary={subsidiary}
           />
 
+          {isError && (
+            <OsdsMessage className="mt-5" color={ODS_THEME_COLOR_INTENT.error}>
+              {t('account-disable-2fa-create-form-error-message-send-document')}
+            </OsdsMessage>
+          )}
           <OsdsButton
             type={ODS_BUTTON_TYPE.submit}
             color={ODS_THEME_COLOR_INTENT.primary}
-            className="w-1/3 mt-10"
+            className="mt-5"
+            inline
             disabled={!(formState.isValid && isAnyFileSelected) || undefined}
           >
             {t('account-disable-2fa-create-form-submit')}
           </OsdsButton>
         </>
       )}
+
+      {showConfirmModal && (
+        <ConfirmModal
+          isPending={isPending}
+          onClose={() => setShowConfirmModal(false)}
+          onValidate={() => {
+            mutate({ files });
+          }}
+        />
+      )}
+      {showSuccessModal && <SuccessModal ovhHomePageHref={ovhHomePageHref} />}
     </form>
   );
 };
