+++ packages/manager/apps/procedures/src/public/translations/account-disable-2fa/Messages_fr_FR.json
@@ -38,5 +38,16 @@
   "account-disable-2fa-session-modal-expired-auth-button": "Retour à l'authentification",
 
   "account-disable-2fa-session-modal-warning-message": "Votre session va bientôt expirer",
-  "account-disable-2fa-session-modal-warning-ok-button": "Ok"
+  "account-disable-2fa-session-modal-warning-ok-button": "Ok",
+
+  "account-disable-2fa-confirm-modal-yes": "Oui",
+  "account-disable-2fa-confirm-modal-no": "Non",
+  "account-disable-2fa-create-form-confirm-modal-send-document-title": "Envoyer mes documents",
+  "account-disable-2fa-create-form-confirm-modal-send-document-description-insure": "Veuillez-vous assurer que tous vos documents sont corrects et lisibles avant envoi. En cas de documents non valides, cette procédure sera annulée et vous devrez en effectuer une nouvelle.",
+  "account-disable-2fa-create-form-confirm-modal-send-document-description-confirm": "Confirmez-vous avoir ajouté l'ensemble des documents demandés ?",
+
+  "account-disable-2fa-create-form-success-modal-send-document-title": "Merci, nous avons bien reçu tous vos documents !",
+  "account-disable-2fa-create-form-success-modal-send-document-description": "Ils vont être soumis à notre équipe pour validation. Notre équipe support vous répond dans les 72 heures ! Vous serez notifié de leur validation par email, ou vous serez contacté si nous avons besoin de plus d'informations.",
+  "account-disable-2fa-success-modal-back-home": "Retour à la page d'accueil",
+  "account-disable-2fa-create-form-error-message-send-document": "Vos documents n’ont pas été correctement envoyés. Veuillez relancer l’envoi de vos documents."
 }
